Fix some bugs, add support for declaring array variables.
git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile/Languages/ObjC2JS@6847 7b3f36aa-e6db-49c1-8b4c-66eac1997e64

+++ Test/runtime.js
@@ -31,8 +31,8 @@ function pointerCastTo(obj, integer, signed, size)
 			switch (size)
 			{
 				case 8:  typedBuffer = Int8Array(buffer);
-				case 16:  typedBuffer = Int8Array(buffer);
-				case 32:  typedBuffer = Int8Array(buffer);
+				case 16:  typedBuffer = Int16Array(buffer);
+				case 32:  typedBuffer = Int32Array(buffer);
 			}
 		}
 		else
@@ -40,8 +40,8 @@ function pointerCastTo(obj, integer, signed, size)
 			switch (size)
 			{
 				case 8:  typedBuffer = Uint8Array(buffer);
-				case 16:  typedBuffer = Uint8Array(buffer);
-				case 32:  typedBuffer = Uint8Array(buffer);
+				case 16:  typedBuffer = Uint16Array(buffer);
+				case 32:  typedBuffer = Uint32Array(buffer);
 			}
 		}
 	}
@@ -53,7 +53,7 @@ function pointerCastTo(obj, integer, signed, size)
 		}
 		else
 		{
-			typedBuffer = Float32Array(buffer);
+			typedBuffer = Float64Array(buffer);
 		}
 	}
 	typedBuffer.buffer = buffer;
